Rename config command to configure
